refactor: address daily quality findings for 2026-08-08
Extract browser-use chrome helpers, host-window pure matching for AppShot
host capture, and Desktop Use engine progress bar to shrink day-grown hubs.
